moved skills.js
